Project Bootstrap
Since we will be working with Redis, it's a good idea to have an
isolated Redis environment that we can use for development purposes.
This allows us to have better control over Redis while we experiment
with different ideas. It also keep the Redis source close by so we can
take a look at implementation details while we write our adapter.

Signed-off-by: Aaron Bedra <aaron@aaronbedra.com>