Added initial exploration
I'm building this from scratch and using my org-mode outline to store
the steps as I do it. This will allow me to track the steps as I go and
the slides should pretty much build themselves.

The initial idea of producing the commands to Redis and sending data
across the wire is complete. Next up is processing the data that is
returned from Redis. This should be preceeded by an introduction to
Polymorphism in Clojure.

Signed-off-by: Aaron Bedra <aaron@aaronbedra.com>