Added most of the code examples for the talk
This adds most of the code examples from the talk. I am still not sure
how much time this will take up, so I am going to stop here for a bit
and turn the outline into slides to get a better feel on time. If it is
enough as it is I will find a good way to wrap up the api usage and
closing ideas. If it still needs more time I will add the examples for
keeping track of the commands programatically.

Signed-off-by: Aaron Bedra <aaron@aaronbedra.com>