Outline updates after creating initial slide deck
I use org-mode to export code snippets to html so I can paste the
results into keynote when creating slideshows. This commit updates the
outline a bit to add some things I needed to paste in. It also adds the
html file so I have it to copy out of going forward.

Signed-off-by: Aaron Bedra <aaron@aaronbedra.com>